Rename conftest.py to avoid collisions.
