+++ tests/unit/plugins/modules/cloud/xenserver/xenserver_common.py
