+++ tests/unit/plugins/modules/packaging/os/rhn_conftest.py
