+++ tests/unit/plugins/modules/packaging/os/test_rhn_channel.py
@@ -10,6 +10,8 @@ import json
 
 from ansible_collections.community.general.plugins.modules.packaging.os import rhn_channel
 
+from .rhn_conftest import mock_request
+
 import pytest
