+++ tests/unit/plugins/modules/remote_management/oneview/test_oneview_datacenter_info.py
@@ -7,8 +7,10 @@ __metaclass__ = type
 
 import pytest
 
-from ansible_collections.community.general.plugins.modules.remote_management.oneview.oneview_datacenter_info import DatacenterInfoModule
 from .hpe_test_utils import FactsParamsTest
+from .oneview_conftest import mock_ov_client, mock_ansible_module
+
+from ansible_collections.community.general.plugins.modules.remote_management.oneview.oneview_datacenter_info import DatacenterInfoModule
 
 PARAMS_GET_CONNECTED = dict(
     config='config.json',
