+++ tests/unit/plugins/modules/remote_management/oneview/test_oneview_enclosure_info.py
@@ -5,9 +5,11 @@
 from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
+from .hpe_test_utils import FactsParamsTestCase
+from .oneview_conftest import mock_ov_client, mock_ansible_module
+
 from ansible_collections.community.general.tests.unit.compat import unittest
 from ansible_collections.community.general.plugins.modules.remote_management.oneview.oneview_enclosure_info import EnclosureInfoModule
-from .hpe_test_utils import FactsParamsTestCase
 
 
 ERROR_MSG = 'Fake message error'
