+++ tests/unit/plugins/modules/remote_management/oneview/test_oneview_ethernet_network.py
@@ -11,6 +11,7 @@ import yaml
 
 from ansible_collections.community.general.tests.unit.compat import unittest, mock
 from .oneview_module_loader import EthernetNetworkModule
+from .oneview_conftest import mock_ov_client, mock_ansible_module
 from .hpe_test_utils import OneViewBaseTestCase
 
 FAKE_MSG_ERROR = 'Fake message error'
