+++ tests/unit/plugins/modules/remote_management/oneview/test_oneview_fcoe_network.py
@@ -9,6 +9,7 @@ __metaclass__ = type
 
 from ansible_collections.community.general.tests.unit.compat import unittest
 from .oneview_module_loader import FcoeNetworkModule
+from .oneview_conftest import mock_ov_client, mock_ansible_module
 from .hpe_test_utils import OneViewBaseTestCase
 
 FAKE_MSG_ERROR = 'Fake message error'
