+++ tests/unit/plugins/modules/remote_management/oneview/test_oneview_logical_interconnect_group_info.py
@@ -6,10 +6,11 @@ from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
 from ansible_collections.community.general.tests.unit.compat import unittest
+from .hpe_test_utils import FactsParamsTestCase
+from .oneview_conftest import mock_ov_client, mock_ansible_module
 from ansible_collections.community.general.plugins.modules.remote_management.oneview.oneview_logical_interconnect_group_info import (
     LogicalInterconnectGroupInfoModule
 )
-from .hpe_test_utils import FactsParamsTestCase
 
 
 ERROR_MSG = 'Fake message error'
