+++ tests/unit/plugins/modules/remote_management/oneview/test_oneview_network_set.py
@@ -8,6 +8,7 @@ __metaclass__ = type
 from ansible_collections.community.general.tests.unit.compat import unittest, mock
 from .hpe_test_utils import OneViewBaseTestCase
 from .oneview_module_loader import NetworkSetModule
+from .oneview_conftest import mock_ov_client, mock_ansible_module
 
 FAKE_MSG_ERROR = 'Fake message error'
