+++ tests/unit/plugins/modules/remote_management/oneview/test_oneview_network_set_info.py
@@ -7,6 +7,7 @@ __metaclass__ = type
 
 from ansible_collections.community.general.tests.unit.compat import unittest
 from .oneview_module_loader import NetworkSetInfoModule
+from .oneview_conftest import mock_ov_client, mock_ansible_module
 from .hpe_test_utils import FactsParamsTestCase
 
 ERROR_MSG = 'Fake message error'
